Add Redux state management for community forum feature
This commit adds Redux state management for the community forum feature, including:

- Adding the communityForumReducer to the store
- Adding actions for adding a post and getting posts in the communitySlice
- Using the useSelector and useDispatch hooks to access and dispatch actions related to the community forum feature
- Updating the CommunityForum page to use the new Redux state management

Additionally, the AppointmentCard component has been updated to use the useSelector hook to access the user object and display the user's name instead of a hardcoded name.

Finally, the layout.tsx file has been updated to include the ReduxProvider component, which wraps the entire application and provides the Redux store to all components.